Refactor checker_service from torrent checker and update tests
